+++ tests/test_dvc.py
@@ -4,16 +4,11 @@ import pytest
 from dvc.exceptions import DvcException
 from dvc.repo import Repo
 from dvc.scm import NoSCM
-from PIL import Image
-from ruamel.yaml import YAML
 from scmrepo.git import Git
 
 from dvclive import Live
-from dvclive.dvc import get_dvc_repo, make_dvcyaml
+from dvclive.dvc import get_dvc_repo
 from dvclive.env import DVC_EXP_BASELINE_REV, DVC_EXP_NAME
-from dvclive.serialize import dump_yaml, load_yaml
-
-YAML_LOADER = YAML(typ="safe")
 
 
 def test_get_dvc_repo(tmp_dir):
@@ -28,432 +23,6 @@ def test_get_dvc_repo_subdir(tmp_dir):
     subdir.mkdir()
     os.chdir(subdir)
     assert get_dvc_repo().root_dir == str(tmp_dir)
-
-
-def test_make_dvcyaml_empty(tmp_dir):
-    live = Live(dvcyaml="dvc.yaml")
-    make_dvcyaml(live)
-
-    assert load_yaml(live.dvc_file) == {}
-
-
-def test_make_dvcyaml_param(tmp_dir):
-    live = Live(dvcyaml="dvc.yaml")
-    live.log_param("foo", 1)
-    make_dvcyaml(live)
-
-    assert load_yaml(live.dvc_file) == {
-        "params": ["dvclive/params.yaml"],
-    }
-
-
-def test_make_dvcyaml_metrics(tmp_dir):
-    live = Live(dvcyaml="dvc.yaml")
-    live.log_metric("bar", 2)
-    make_dvcyaml(live)
-
-    assert load_yaml(live.dvc_file) == {
-        "metrics": ["dvclive/metrics.json"],
-        "plots": [{"dvclive/plots/metrics": {"x": "step"}}],
-    }
-
-
-def test_make_dvcyaml_metrics_no_plots(tmp_dir):
-    live = Live(dvcyaml="dvc.yaml")
-    live.log_metric("bar", 2, plot=False)
-    make_dvcyaml(live)
-
-    assert load_yaml(live.dvc_file) == {
-        "metrics": ["dvclive/metrics.json"],
-    }
-
-
-def test_make_dvcyaml_summary(tmp_dir):
-    live = Live(dvcyaml="dvc.yaml")
-    live.summary["bar"] = 2
-    make_dvcyaml(live)
-
-    assert load_yaml(live.dvc_file) == {
-        "metrics": ["dvclive/metrics.json"],
-    }
-
-
-def test_make_dvcyaml_all_plots(tmp_dir):
-    live = Live(dvcyaml="dvc.yaml")
-    live.log_param("foo", 1)
-    live.log_metric("bar", 2)
-    live.log_image("img.png", Image.new("RGB", (10, 10), (250, 250, 250)))
-    live.log_sklearn_plot("confusion_matrix", [0, 0, 1, 1], [0, 1, 1, 0])
-    live.log_sklearn_plot(
-        "confusion_matrix",
-        [0, 0, 1, 1],
-        [0, 1, 1, 0],
-        name="confusion_matrix_normalized",
-        normalized=True,
-    )
-    live.log_sklearn_plot("roc", [0, 0, 1, 1], [0.0, 0.5, 0.5, 0.0], "custom_name_roc")
-    make_dvcyaml(live)
-
-    assert load_yaml(live.dvc_file) == {
-        "metrics": ["dvclive/metrics.json"],
-        "params": ["dvclive/params.yaml"],
-        "plots": [
-            {"dvclive/plots/metrics": {"x": "step"}},
-            "dvclive/plots/images",
-            {
-                "dvclive/plots/sklearn/confusion_matrix.json": {
-                    "template": "confusion",
-                    "x": "actual",
-                    "y": "predicted",
-                    "title": "Confusion Matrix",
-                    "x_label": "True Label",
-                    "y_label": "Predicted Label",
-                },
-            },
-            {
-                "dvclive/plots/sklearn/confusion_matrix_normalized.json": {
-                    "template": "confusion_normalized",
-                    "title": "Confusion Matrix",
-                    "x": "actual",
-                    "x_label": "True Label",
-                    "y": "predicted",
-                    "y_label": "Predicted Label",
-                }
-            },
-            {
-                "dvclive/plots/sklearn/custom_name_roc.json": {
-                    "template": "simple",
-                    "x": "fpr",
-                    "y": "tpr",
-                    "title": "Receiver operating characteristic (ROC)",
-                    "x_label": "False Positive Rate",
-                    "y_label": "True Positive Rate",
-                }
-            },
-        ],
-    }
-
-
-def test_make_dvcyaml_relpath(tmp_dir, mocked_dvc_repo):
-    (tmp_dir / "model.pth").touch()
-    live = Live(dvcyaml="dir/dvc.yaml")
-    live.log_metric("foo", 1)
-    live.log_artifact("model.pth", type="model")
-    make_dvcyaml(live)
-
-    assert load_yaml(live.dvc_file) == {
-        "metrics": ["../dvclive/metrics.json"],
-        "plots": [{"../dvclive/plots/metrics": {"x": "step"}}],
-        "artifacts": {
-            "model": {"path": "../model.pth", "type": "model"},
-        },
-    }
-
-
-def test_make_dvcyaml_update_stages(tmp_dir):
-    orig_yaml = {
-        "stages": {"train": {"cmd": "train.py"}},
-    }
-
-    updated_yaml = {
-        "stages": {"train": {"cmd": "train.py"}},
-        "metrics": ["dvclive/metrics.json"],
-        "plots": [
-            {"dvclive/plots/metrics": {"x": "step"}},
-        ],
-    }
-
-    dump_yaml(orig_yaml, "dvc.yaml")
-
-    live = Live(dvcyaml="dvc.yaml")
-    live.log_metric("foo", 2)
-    make_dvcyaml(live)
-
-    assert load_yaml(live.dvc_file) == updated_yaml
-
-
-def test_make_dvcyaml_update_keep_custom_plots(tmp_dir):
-    orig_yaml = {
-        "plots": [
-            {
-                "dvclive_custom_title": {
-                    "x": "step",
-                    "y": {"dvclive/metrics/acc.tsv": "acc"},
-                    "title": "Accuracy",
-                }
-            }
-        ],
-    }
-
-    updated_yaml = {
-        "metrics": ["dvclive/metrics.json"],
-        "plots": [
-            {
-                "dvclive_custom_title": {
-                    "x": "step",
-                    "y": {"dvclive/metrics/acc.tsv": "acc"},
-                    "title": "Accuracy",
-                }
-            },
-            {"dvclive/plots/metrics": {"x": "step"}},
-        ],
-    }
-
-    dump_yaml(orig_yaml, "dvc.yaml")
-
-    live = Live(dvcyaml="dvc.yaml")
-    live.log_metric("foo", 2)
-    make_dvcyaml(live)
-
-    assert load_yaml(live.dvc_file) == updated_yaml
-
-
-def test_make_dvcyaml_update_drop_extra_sections(tmp_dir):
-    orig_yaml = {
-        "params": ["dvclive/params.yaml"],
-    }
-
-    updated_yaml = {
-        "metrics": ["dvclive/metrics.json"],
-        "plots": [
-            {"dvclive/plots/metrics": {"x": "step"}},
-        ],
-    }
-
-    dump_yaml(orig_yaml, "dvc.yaml")
-
-    live = Live(dvcyaml="dvc.yaml")
-    live.log_metric("foo", 2)
-    make_dvcyaml(live)
-
-    assert load_yaml(live.dvc_file) == updated_yaml
-
-
-def test_make_dvcyaml_update_drop_unlogged_plots(tmp_dir):
-    orig_yaml = {
-        "plots": [
-            "dvclive/plots/images",
-        ]
-    }
-
-    updated_yaml = {
-        "metrics": ["dvclive/metrics.json"],
-        "plots": [
-            {"dvclive/plots/metrics": {"x": "step"}},
-        ],
-    }
-
-    dump_yaml(orig_yaml, "dvc.yaml")
-
-    live = Live(dvcyaml="dvc.yaml")
-    live.log_metric("foo", 2)
-    make_dvcyaml(live)
-
-    assert load_yaml(live.dvc_file) == updated_yaml
-
-
-def test_make_dvcyaml_update_plot_props(tmp_dir):
-    orig_yaml = {
-        "plots": [
-            {"dvclive/plots/metrics": {"x": "step", "y": "foo"}},
-        ]
-    }
-
-    updated_yaml = {
-        "metrics": ["dvclive/metrics.json"],
-        "plots": [
-            {"dvclive/plots/metrics": {"x": "step"}},
-        ],
-    }
-
-    dump_yaml(orig_yaml, "dvc.yaml")
-
-    live = Live(dvcyaml="dvc.yaml")
-    live.log_metric("foo", 2)
-    make_dvcyaml(live)
-
-    assert load_yaml(live.dvc_file) == updated_yaml
-
-
-def test_make_dvcyaml_keep_custom_plots(tmp_dir):
-    orig_yaml = {
-        "plots": [
-            {
-                "custom": {
-                    "x": "step",
-                    "y": {"dvclive/plots/metrics": "foo"},
-                    "title": "custom",
-                }
-            },
-        ],
-    }
-
-    updated_yaml = {
-        "metrics": ["dvclive/metrics.json"],
-        "plots": [
-            {
-                "custom": {
-                    "x": "step",
-                    "y": {"dvclive/plots/metrics": "foo"},
-                    "title": "custom",
-                }
-            },
-            {"dvclive/plots/metrics": {"x": "step"}},
-        ],
-    }
-
-    dump_yaml(orig_yaml, "dvc.yaml")
-
-    live = Live(dvcyaml="dvc.yaml")
-    live.log_metric("foo", 2)
-    make_dvcyaml(live)
-
-    assert load_yaml(live.dvc_file) == updated_yaml
-
-
-def test_make_dvcyaml_update_artifact_props(tmp_dir, mocked_dvc_repo):
-    orig_yaml = {
-        "artifacts": {
-            "model": {"path": "model.pth", "type": "model", "desc": "best model"},
-        },
-    }
-
-    updated_yaml = {
-        "artifacts": {
-            "model": {"path": "dvclive/artifacts/model.pth", "type": "model"},
-        },
-    }
-
-    dump_yaml(orig_yaml, "dvc.yaml")
-    (tmp_dir / "model.pth").touch()
-
-    live = Live(dvcyaml="dvc.yaml")
-    live.log_artifact("model.pth", type="model", copy=True)
-    make_dvcyaml(live)
-
-    assert load_yaml(live.dvc_file) == updated_yaml
-
-
-def test_make_dvcyaml_update_drop_duplicate_artifact(tmp_dir, mocked_dvc_repo):
-    orig_yaml = {
-        "artifacts": {
-            "duplicate": {"path": "dvclive/artifacts/model.pth"},
-        },
-    }
-
-    updated_yaml = {
-        "artifacts": {
-            "model": {"path": "dvclive/artifacts/model.pth", "type": "model"},
-        },
-    }
-
-    dump_yaml(orig_yaml, "dvc.yaml")
-    (tmp_dir / "model.pth").touch()
-
-    live = Live(dvcyaml="dvc.yaml")
-    live.log_artifact("model.pth", type="model", copy=True)
-    make_dvcyaml(live)
-
-    assert load_yaml(live.dvc_file) == updated_yaml
-
-
-def test_make_dvcyaml_update_keep_extra_artifact(tmp_dir, mocked_dvc_repo):
-    orig_yaml = {
-        "artifacts": {
-            "data": {"path": "data.csv", "desc": "source data"},
-        },
-    }
-
-    updated_yaml = {
-        "artifacts": {
-            "model": {"path": "dvclive/artifacts/model.pth", "type": "model"},
-            "data": {"path": "data.csv", "desc": "source data"},
-        },
-    }
-
-    dump_yaml(orig_yaml, "dvc.yaml")
-    (tmp_dir / "model.pth").touch()
-
-    live = Live(dvcyaml="dvc.yaml")
-    live.log_artifact("model.pth", type="model", copy=True)
-    make_dvcyaml(live)
-
-    assert load_yaml(live.dvc_file) == updated_yaml
-
-
-def test_make_dvcyaml_update_all(tmp_dir, mocked_dvc_repo):
-    orig_yaml = {
-        "stages": {"train": {"cmd": "train.py"}},
-        "metrics": [
-            "dvclive/metrics.json",
-            "dvclive/metrics.yaml",
-            "other/metrics.json",
-        ],
-        "params": ["dvclive/params.yaml"],
-        "plots": [
-            {"dvclive/plots/metrics": {"x": "step", "y": "foo"}},
-            "dvclive/plots/images",
-            "other/plots",
-            {
-                "custom": {
-                    "x": "step",
-                    "y": {"dvclive/plots/metrics": "foo"},
-                    "title": "custom",
-                }
-            },
-            {
-                "dvclive/plots/sklearn/confusion_matrix.json": {
-                    "template": "confusion",
-                    "x": "actual",
-                    "y": "predicted",
-                    "title": "Confusion Matrix",
-                    "x_label": "True Label",
-                    "y_label": "Predicted Label",
-                },
-            },
-        ],
-        "artifacts": {
-            "model": {"path": "dvclive/artifacts/model.pth", "type": "model"},
-            "duplicate": {"path": "dvclive/artifacts/model.pth"},
-            "data": {"path": "data.csv", "desc": "source data"},
-            "other": {"path": "other.pth"},
-        },
-    }
-
-    updated_yaml = {
-        "stages": {"train": {"cmd": "train.py"}},
-        "metrics": ["other/metrics.json", "dvclive/metrics.json"],
-        "plots": [
-            "other/plots",
-            {
-                "custom": {
-                    "x": "step",
-                    "y": {"dvclive/plots/metrics": "foo"},
-                    "title": "custom",
-                }
-            },
-            {"dvclive/plots/metrics": {"x": "step"}},
-            "dvclive/plots/images",
-        ],
-        "artifacts": {
-            "model": {"path": "dvclive/artifacts/model.pth", "type": "model"},
-            "data": {"path": "data.csv", "desc": "source data"},
-            "other": {"path": "other.pth"},
-        },
-    }
-
-    dump_yaml(orig_yaml, "dvc.yaml")
-    (tmp_dir / "model.pth").touch()
-    (tmp_dir / "data.csv").touch()
-
-    live = Live(dvcyaml="dvc.yaml")
-    live.log_metric("foo", 2)
-    live.log_image("img.png", Image.new("RGB", (10, 10), (250, 250, 250)))
-    live.log_artifact("model.pth", type="model", copy=True)
-    live.log_artifact("data.csv", desc="source data")
-    make_dvcyaml(live)
-
-    assert load_yaml(live.dvc_file) == updated_yaml
 
 
 @pytest.mark.parametrize("save", [True, False])
@@ -508,26 +77,6 @@ def test_exp_save_run_on_dvc_repro(tmp_dir, mocker):
     dvc_repo.experiments.save.assert_called_with(
         name=live._exp_name, include_untracked=[live.dir], force=True, message=None
     )
-
-
-@pytest.mark.parametrize("dvcyaml", [True, False])
-def test_dvcyaml_on_next_step(tmp_dir, dvcyaml, mocked_dvc_repo):
-    live = Live(dvcyaml=dvcyaml)
-    live.next_step()
-    if dvcyaml:
-        assert (tmp_dir / live.dvc_file).exists()
-    else:
-        assert not (tmp_dir / live.dvc_file).exists()
-
-
-@pytest.mark.parametrize("dvcyaml", [True, False])
-def test_dvcyaml_on_end(tmp_dir, dvcyaml, mocked_dvc_repo):
-    live = Live(dvcyaml=dvcyaml)
-    live.end()
-    if dvcyaml:
-        assert (tmp_dir / live.dvc_file).exists()
-    else:
-        assert not (tmp_dir / live.dvc_file).exists()
 
 
 def test_exp_save_with_dvc_files(tmp_dir, mocker):
@@ -608,21 +157,6 @@ def test_errors_on_git_add_are_catched(tmp_dir, mocked_dvc_repo, monkeypatch):
         live.summary["foo"] = 1
 
 
-def test_make_dvcyaml_idempotent(tmp_dir, mocked_dvc_repo):
-    (tmp_dir / "model.pth").touch()
-
-    with Live() as live:
-        live.log_artifact("model.pth", type="model")
-
-    live.make_dvcyaml()
-
-    assert load_yaml(live.dvc_file) == {
-        "artifacts": {
-            "model": {"path": "model.pth", "type": "model"},
-        }
-    }
-
-
 def test_exp_save_message(tmp_dir, mocked_dvc_repo):
     live = Live(exp_message="Custom message")
     live.end()
@@ -644,22 +178,3 @@ def test_no_scm_repo(tmp_dir, mocker):
 
         live = Live()
         assert live._save_dvc_exp is False
-
-
-@pytest.mark.parametrize("dvcyaml", [True, False, "dvclive/dvc.yaml"])
-def test_warn_on_dvcyaml_output_overlap(tmp_dir, mocker, mocked_dvc_repo, dvcyaml):
-    logger = mocker.patch("dvclive.live.logger")
-    dvc_stage = mocker.MagicMock()
-    dvc_stage.addressing = "train"
-    dvc_out = mocker.MagicMock()
-    dvc_out.fs_path = tmp_dir / "dvclive"
-    dvc_stage.outs = [dvc_out]
-    mocked_dvc_repo.index.stages = [dvc_stage]
-    live = Live(dvcyaml=dvcyaml)
-
-    if dvcyaml == "dvclive/dvc.yaml":
-        msg = f"'{live.dvc_file}' is in outputs of stage 'train'.\n"
-        msg += "Remove it from outputs to make DVCLive work as expected."
-        logger.warning.assert_called_with(msg)
-    else:
-        logger.warning.assert_not_called()
